PowerPoint Presentation
Our project describes IoT-based smart Bio-Tractor. The main motive of our project is to reduce the landfill problems and to protect our environment from pollution. More than 7.5 billion tons of solid wastes are dumped every year. This causes lots of problems such as soil erosion, various diseases and causes pollution. This bio tractor is used to destroy the wastes using incinerating process. When the hot and cold body reacts together electricity is generated using Peltier module, So in hot junction, we will be burning the wastes and the air pollutant is captured and turned into ink from the presence of carbon material in the air pollutant using kaalink device and in cold junction, we will be using water or any other coolant. In cold junction, we will have an ultrasonic sensor when shows the level of water or any other coolant in the IBM cloud. The solar panel is fixed in the top of the bio tractor which produces additional electricity. Electricity from the solar panel and Peltier module is used to charge the battery-based bio tractor. This tractor can be remotely controlled using robo vehicle application in all direction which will be helpful for the farmers.